requirements.txt
flask
python-dotenv
authlib
requests
dataclasses-json
PyJWT